refactor(backend): share one file_name_of helper between transfer commands
Both commands/files.rs (sftp) and commands/transfer.rs (ftp) had a
byte-for-byte identical private file_name_of helper after #1573 and
#1594 converged on the same rule: a transfer row's name is the basename
of the path it displays. Extract a single pub file_name_of into
utils/fs.rs (the existing home for small filesystem predicates) and have
both call sites import it. The ftp import is cfg-gated so it does not
warn as unused when the ftp feature is off; the sftp call sites keep it
live regardless. The two identical unit-test modules are folded into one
set against the shared helper.

Closes #1612
